cleanup JS files
